refactor: improve markup and structure to better match the provided design layout
