+++ src/features/chats/index.ts
@@ -1,1 +1,6 @@
 export * from './ui/chat-sidebar'
+export * from './hooks/use-chat-socket'
+export * from './ui/chat-list'
+export * from './ui/chat-sidebar'
+export * from './ui/create-chat-modal'
+export * from './ui/messages-list'
