--- src/features/chats/ui/index.ts
@@ -1,2 +1,0 @@
-export * from './chat-list'
-export * from './chat-sidebar'
